Fix bugs about GrpcCoordinator

+++ examples/src/main/resources/example.sql
@@ -24,7 +24,7 @@ CREATE TABLE `author_posts`(
   `first_name` varchar(50) COLLATE utf8_unicode_ci NOT NULL,
   `last_name` varchar(50) COLLATE utf8_unicode_ci NOT NULL,
   `email` varchar(100) COLLATE utf8_unicode_ci NOT NULL,
-  `posts` int(11) NOT NULL,
+  `posts` bigint(32) NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB AUTO_INCREMENT=101 DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
